pyproject.toml
[tool.poetry]
name = "bewise"
version = "0.1.0"
description = "test task"
authors = ["NicholasNartikoev <nartikoevnicholas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.22"
fastapi = "^0.103.2"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.9"
pydantic-config = "^0.2.3"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
